requirements.txt
discord.py[voice]
dnspython
motor
python-dotenv
pytz